`GetStaffs`関数をusers.goに移動
